Revert "TESTFIX: Updated all diff'ed against files because of modified precision (float64)."
dds_basetype is now back to old tf.float32.

- we go back to the old files in principle. However, as columns have changed
  and also the output precision a direct comparison (within the scope of
  commits) is not possible. Nonetheless, we DO obtain the same values as
  (long) before.
- TESTFIX: Changed all diff'ed against files.
- TESTFIX: hmc hidden regression tests sees slight fluctuation on each run of
  unknown origin at 3e-7 relative precision on loss, total_energy, and old_....
- TESTFIX: Pior lower boundary needs prior force of 2. (instead of 1.) to keep
  dofs away from boundary.
- TESTFIX: Userguide test on lossfunction grid tests now _relative_ 1e-5
  precision.
- TESTFIX: Userguide test on optimize has different final loss value (basically
  the same change as to DataOptimizer twoclusters test).

+++ src/DataDrivenSampler/models/basetype.py
@@ -1,3 +1,3 @@
 import tensorflow as tf
 
-dds_basetype = tf.float64
+dds_basetype = tf.float32
